kimendar/requirements.txt
python-telegram-bot==11.1.0
python-redmine==2.1.1